ui-fixes

+++ client/src/components/Item.jsx
@@ -3,13 +3,16 @@ import { Link, useNavigate } from "react-router-dom";
 import { ItemContext } from "../context/ItemContext";
 
 const Item = ({ id, name, price, stock }) => {
-  const navigate = useNavigate();
   const { getItems } = useContext(ItemContext);
   const deleteItem = async () => {
-    const response = await fetch(`${import.meta.env.VITE_SERVER}/item/${id}`, {
-      method: "delete",
-    });
-    getItems();
+    if (confirm("Are you sure?")) {
+      await fetch(`${import.meta.env.VITE_SERVER}/item/${id}`, {
+        method: "delete",
+      });
+      getItems();
+    } else {
+      return;
+    }
   };
 
   // const updateStock = async (action) => {
@@ -31,16 +34,25 @@ const Item = ({ id, name, price, stock }) => {
   //   getItems();
   // };
   return (
-    <div className="border-2 border-red-300 p-5 w-[20%] rounded-lg m-2 cursor-pointer">
-      <h1>{name}</h1>
-      <p>Price - {price}</p>
-      <div className="flex gap-3">
+    <div className="border-2 m-auto border-red-300 p-1 w-[95%] rounded-lg my-2 cursor-pointer flex flex-col gap-4 items-center justify-left text-left">
+      <h1 className="text-red-400 font-semibold">{name}</h1>
+      <div className="flex gap-5 items-center">
+        <p>Price - {price}</p>
         <p>Stock - {stock}</p>
-      </div>
-      <Link to={`http://localhost:5173/item/${id}`}>View details</Link>
-      <div className="flex gap-5">
-        <button onClick={deleteItem}>Delete Item</button>
-        <Link to={`/edit/item/${id}`}>Edit Item</Link>
+        <Link
+          className="hover:text-red-300"
+          to={`http://localhost:5173/item/${id}`}
+        >
+          Details
+        </Link>
+        <div className="flex gap-5">
+          <button className="hover:text-red-300" onClick={deleteItem}>
+            Delete
+          </button>
+          <Link className="hover:text-red-300" to={`/edit/item/${id}`}>
+            Edit
+          </Link>
+        </div>
       </div>
     </div>
   );
